Clean up file names
