+++ internal/command/env_collector_fifo_unix_test.go
