+++ internal/command/env_collector_fifo_windows.go
